% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pals.R
\name{fiege_cols}
\alias{fiege_cols}
\title{Function to extract fiege colors as hex codes}
\usage{
fiege_cols(...)
}
\arguments{
\item{...}{Character names of fiege_colors}
}
\description{
Function to extract fiege colors as hex codes
}
