% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pals.R
\name{fiege_pal}
\alias{fiege_pal}
\title{Return function to interpolate a fiege color palette}
\usage{
fiege_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in fiege_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a fiege color palette
}
